Stardate: 03112014
Well, this is what I have right now. Nolan's code is currently in its
own class and I wasn't able to integrate it properly. Or I was, but I
run into problems with the display. Will update again soon. Right now,
working on allowing the black pieces to be controlled by another player
so it can also act as a two player game.